requirements.txt
colorama
requests
whois
dnspython
bs4
psutil
tabulate
scapy
python-whois
speedtest-cli
phonenumbers
ping3
smbprotocol